Updated indentation
Fixed identations. To improve readability in html documents its recomended to keep a precise identation.
To change the subject: is very specific the use of tables in modern html documents. If you're looking to design websites later on I recommend the use of lists and just order everything with CSS. With flexbox or a grid system you can reach the same result with half the effort